rift: Split UVC setup/destroy from start/stop
Preparation for having 1 explicit capture thread and 1 processing thread
instead of having each UVC capture run a separate thread that
fight over a single lock to dispatch USB events, and then process
completed frames in-thread.

The goal is to have a ping-pong capture frame for each sensor camera,
and hand them off to a separate thread for processing so it doesn't
block USB Isoc packet retrieval of the next frame.