rift: Blobwatch improvements
Make x/y position of a blob be floating point, so we can track
sub-pixel positions.

Extend the LED pattern to an int, so we have room to store the
device ID of the LED in the (8..15) bits.

Expand the number of extents and blobs we allow, and permit
smaller 2 pixel wide extents, as the CV1 has many more LEDs
involved, and they are smaller.

Get ready to track a different pixel threshold for CV1, as
it excludes background noise a lot better, and has faint LEDs
that go to lower pixel values.