rift: Track the exposure associated with frames differently.
With multiple sensors, there was a race where if the frame from one sensor finished
arriving and was analysed very quickly, the frame might be released
before the frame(s) finish arriving on other sensor(s). In that
situation, the use_count for the associated delay_slot would hit 0
immediately and the delay slot would be invalidated, preventing any
poses from other sensors being processed.

To avoid that race, increment the use_count for the delay slot whenever
any sensor is capturing a frame for that exposure, and release the slots
if the exposure changes mid-flight (while a frame is arriving).