rift: Rework tracker pose calculations.
Change the way poses are calculated in the tracker
when applying / unapplying the IMU offset and handling
the quirk where the HMD LED model is reversed compared
to the IMU. This fixes a few-cm offset in reported
positions - especially for the controllers.

Also implement conversion for the angular velocity
and projection to generate linear velocity at the
device pose position vs IMU position.