rift: Store each pose observation and reclaim delay slots.
As pose reports from each sensor arrive, store them in an array.
Later, we can defer integrating the reports  perhaps. For now,
if we need a delay slot and there's one that's already received
a pose report, reclaim it. This might mean we miss fewer
frames because we don't have a free delay slot.