rift: Rework frame capture and exposure tracking.
Put all the USB camera operations into a separate object
that the sensor analyses threads are fed from.

Significantly rework how frames are passed around from the
USB sensor device to the analysis threads, and how they
are associated to exposure reports from the HMD.

This will make it easier to support JPEG encoded frames
in the future.

Simplify the frame to exposure association by doing it in
1 step when a frame has fully arrived, instead of needing
operations on a frame_start callback that isn't available
with other capture interfaces like v4l.